Added downloading attachments functionality ISSUE #52
Added a route to support downloading emails from the viewEmail route. Added helper functions in the models.py file to support this downloading (finds user's downloads folder, then creates a new file and writes the contents of the attachment to it). Also slightly edited the viewEmail route to support this functionality.