Added delete confirmation logic (user can recover deleted notes) ISSUE #54
I added a boolean variable to the "Note" model in order to be able to store both deleted, and non-deleted notes. In this new way, a user can view deleted notes, and decide to either recover them, or permanently delete them. This all entailed the creation of 2 routes, and a new template "deleted-notes.html".