Added note importing ability. Also cleaned up forward/reply. ISSUE #54
Added the ability for notes to be imported into email messages from the compose path or the reply/forward paths. Also cleaned up the viewEmail.html file to work more simply (the route for this html file I cleaned up too). viewEmail route no longer uses ForwardReplyForm and instead just a ComposeForm (the forward/reply_flag are accounted for as arguments in the url).